Throttle lastActivity updates to once per second.
The throttling done here should be unnoticeable by end user: suppose the warning is visible on screen and user moves the mouse, and we're concerned there might be ~1s delay until the warning disappears. But, the warning being on the screen means the user has been idle for a good while. So the (now - this.lastActivity < 1000) check would fall through.